+++ examples/jest/CHANGELOG.md
@@ -1,5 +1,12 @@
 # example-jest
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies [3e9dd78]
+  - @firestore-emulator/jest@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes