+++ packages/jest/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @firestore-emulator/jest
 
+## 0.0.5
+
+### Patch Changes
+
+- 3e9dd78: fix sub collections
+- Updated dependencies [3e9dd78]
+  - @firestore-emulator/server@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes