+++ packages/server/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @firestore-emulator/server
 
+## 0.0.5
+
+### Patch Changes
+
+- 3e9dd78: fix sub collections
+
 ## 0.0.4
 
 ### Patch Changes